karol/VHRBluetoothNew (#688)
* changed order of method declarations inside BluetoothAutoConnectService

* Snapshot functionality implemented in BluetoothAutoConnectService

* changed order of method declarations inside BluetoothAutoConnectService

* Snapshot functionality implemented in BluetoothAutoConnectService

* Beginning error handling implementations if dtc or pid requests fail

* Logic for appending dtc to package

* changed order of method declarations inside BluetoothAutoConnectService

* Snapshot functionality implemented in BluetoothAutoConnectService

* Beginning error handling implementations if dtc or pid requests fail

* Logic for appending dtc to package

* Changed scan presenter so that it is compatible with new interface

* Changed scan presenter so that it is compatible with new interface

* Changed scan presenter so that it is compatible with new interface

* Pid observer now retrieves them in a HashMap

* Removed historical data toggle

* fixed time interval for reading pid

* Removed useless funcitonality from BluetoothAutoConnectService

* dtc now returned in hashmap which contains a status

* changed how dtcs are passed to observer

* change to append logic

* terminal rtc time retrieval changed

* terminal rtc time retrieval changed

* terminal rtc time retrieval changed

* terminal rtc time retrieval changed

* terminal rtc time retrieval changed

* rebase fixes

* rebase fixes

* changes to when timers are cancelled

* vin timer

* Changed scan presenter

* terminal rtc time retrieval changed

* rebase fixes

* changed order of method declarations inside BluetoothAutoConnectService

* Snapshot functionality implemented in BluetoothAutoConnectService

* Beginning error handling implementations if dtc or pid requests fail

* changed order of method declarations inside BluetoothAutoConnectService

* Snapshot functionality implemented in BluetoothAutoConnectService

* Beginning error handling implementations if dtc or pid requests fail

* Changed scan presenter so that it is compatible with new interface

* Changed scan presenter so that it is compatible with new interface

* Changed scan presenter so that it is compatible with new interface

* Pid observer now retrieves them in a HashMap

* Removed historical data toggle

* Removed useless funcitonality from BluetoothAutoConnectService

* dtc now returned in hashmap which contains a status

* changed how dtcs are passed to observer

* change to append logic

* terminal rtc time retrieval changed

* terminal rtc time retrieval changed

* rebase fixes

* changes to when timers are cancelled

* vin timer

* rebase master-dev

* changed tripIdRaw to long, and no longer interrupting scan when pids are received inside scan presenter

* changed tripIdRaw to long, and no longer interrupting scan when pids are received inside scan presenter

+++ app/src/pitstop/java/com/pitstop/interactors/other/HandlePidDataUseCaseImpl.java
@@ -135,9 +135,9 @@ public class HandlePidDataUseCaseImpl implements HandlePidDataUseCase {
     }
 
     private Trip215 pidPackageToTrip215Start(PidPackage pidPackage){
-        int tripIdRaw;
+        long tripIdRaw;
         try{
-             tripIdRaw = Integer.valueOf(pidPackage.tripId);
+             tripIdRaw = Long.valueOf(pidPackage.tripId);
         }catch(NumberFormatException e){
             e.printStackTrace();
             tripIdRaw = -1;
